Fix DefaultLocale test that fails in JUnit Pioneer 2.x (#94)
Using the "Hant" variant for Min Man Chinese (language: nan) and
Taiwan (country: TW) fails in JUnit Pioneer 2.0.0 because of breaking
changes.

See the issue here:
https://github.com/junit-pioneer/junit-pioneer/issues/658

Basically, Pioneer changed from using the Locale constructors (which are
deprecated as of Java 19) to the Locale Builder, which is more strict and
checks the variant. Trying to use "Hant" as the variant throws an
IllformedLocaleException. So, used the same language, country, and
variant that Pioneer uses in its own tests: ja, JP, and japanese
respectively.